pulsecore: added reliable PID tracking for (local) clients.
Automatically retrieve and store client credentials if those
are available and enabled. Credentials will be made available
for clients that connect using a socket/protocol family that
is capable of passing credentials (UNIX domain sockets in
practice).

These credentials (PID, user id, and primary group id) can be
used for a number of things, for instance to apply some policy
to clients or object created/owned by clients based on the
user id or group id. Our subsequent patches will use the PID
for implementing a mechanism to make a configurable selection
of whitelisted environment variables available on a per client
basis. A policy enforcement module will then take these into
account when making automatic routing and other decisions.

Note that these credentials have already been queried and
available at the lower communication layers (I/O channels and
pstream). This patch merely propagates them all the way up to
the client.

Currently credentials will not be be available until the first
protocol message is received. If it turns out that credentials
are needed immediately after connection establishment, it is
easy to extend pstream to explictly query the credentials over
its I/O channel instead of waiting for the first packet.

Change-Id: I8ea306761e0b901a1bdf3ec9203920458162d670
Signed-off-by: Krisztian Litkey <kli@iki.fi>