pulsecore: added pa_client_getenv.
Added pa_client_getenv, the core function to retrieve
environment variables from clients' environment. The
actual code to retrieve those variables is not in core.
If the functionality is needed a module can be loaded
to provide an implementation.

The default module-expose-environment provides one such
implementation. If it is available but not loaded, it
will be demand-loaded when the first call is made to
pa_client_getenv.

If no such module is available and none of the loaded
modules provides an implementation, it will not be
possible to retrieve client environment variables.

Alternative implementations with more restrictions or
with a different logic are easily possible by loading
a different module that hooks itself up to the same
minimalistic core infra.

Change-Id: I60e784268e8bcbde5d2f9a3f6aefd48b2254badb
Signed-off-by: Krisztian Litkey <kli@iki.fi>